requirements.txt
numpy
pandas
matplotlib
scikit-learn
torchvision
torchaudio
jupyterlab
tqdm
h5py
wandb

